fix(runtime): keep launch labels process-local
